Updated Portfolio grid to be responsive.
The Portfolio page resized images by scaling them down, and did not
maintain the aspect ratio. Additionally, on smaller screens (mobile
screens) the images were scaled even smaller.

Set image width to 100% so that images are scaled while maintaining
aspect ration. Implemented responsiveness for the Portfolio album by placing it within a responsive div. Edited the styles for portfolio items so that they aren't scaled to 33% of the space available.