Specify exact version of Devise
Hardcode devise version to 4.4.1 to avoid issue with secret key, issue
that only manifests in version 4.4.3.

Deployment to Heroku failed during asset precompilation, with an error
indicating that devise couldn't load the secret_key. The secret_key
should pull from the secret_key_base which is read from the environment.
However a bug in devise 4.4.3 seems to prevent this.